+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+19",
+  "version": "0.8.5+20",
   "private": false,
   "scripts": {
     "dev": "next dev",
